Update SSH sample
