+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/ssh/index.html
@@ -65,6 +65,5 @@
     <script type="text/javascript" src="./authConfig.js"></script>
     <script type="text/javascript" src="./ui.js"></script>  
     <script type="text/javascript" src="./auth.js"></script>
-    <script type="text/javascript" src="./resource.js"></script>
   </body>
 </html>
